Dialogs and Toasts

+++ conFusion/src/app/favorites/favorites.component.ts
@@ -5,7 +5,8 @@ import { ListViewEventData, RadListView } from 'nativescript-telerik-ui/listview
 import { RadListViewComponent } from 'nativescript-telerik-ui/listview/angular';
 import { ObservableArray } from 'tns-core-modules/data/observable-array';
 import { DrawerPage } from '../shared/drawer/drawer.page';
-
+import { confirm } from "ui/dialogs";
+import { Toasty } from "nativescript-toasty";
 @Component({
     selector: 'app-favorites',
     moduleId: module.id,
@@ -32,9 +33,29 @@ export class FavoritesComponent extends DrawerPage implements OnInit {
     }
 
     deleteFavorite(id: string) {
-        this.favoriteService.deleteFavorite(id)
-            .subscribe(avorites => this.favorites = new ObservableArray(favorites),
-            errmess => this.errMess = errmess);
+        console.log('delete', id);
+
+        let options = {
+            title: "Confirm Delete",
+            message: 'Do you want to delete Dish '+ id,
+            okButtonText: "Yes",
+            cancelButtonText: "No",
+            neutralButtonText: "Cancel"
+        }; 
+
+        confirm(options).then((result: boolean) => {
+            if(result) {
+                this.favorites = null;
+                this.favoriteService.deleteFavorite(id)
+                    .subscribe(favorites => {
+                        const toast = new Toasty("Deleted Dish " + id, "short", "bottom");
+                        toast.show();
+                        this.favorites = new ObservableArray(favorites);
+                    },
+                    errmess => this.errMess = errmess);
+            } else {
+                console.log('Delete cancelled');
+            }});
     }
 
     public onCellSwiping(args: ListViewEventData) {
